Refactor Home component to use custom hook for media fetching
